requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flasgger==0.9.7.1
pydantic==2.5
python-dotenv==1.0.0